requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==7.1.2
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
Django==4.2.1
filelock==3.12.0
flake8==6.0.0
fonttools==4.39.4
greenlet==2.0.2
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
mccabe==0.7.0
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.10.0
pycountry==22.3.5
pycountrycodes==0.3.1
pydantic
pyflakes==3.0.1
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
rapidfuzz==2.15.1
requests==2.30.0
rich==9.13.0
simplejson==3.19.1
six==1.16.0
SQLAlchemy==1.4.48
sqlparse==0.4.4
tabulate==0.8.10
task==0.2.5
tomli==2.0.1
tox==3.28.0
typing-extensions
urllib3==2.0.2
virtualenv==20.23.0
celery==5.1.2
click==7.1.2
redis==4.5.5
